requirements.txt

requests
pyyaml
browser-cookie3
python-dotenv
rich
